move `FeedStore` conformance of `RealmFeedStore` into a separate file
